feat: Handle admin verbs better
Instead of scroll through a list of items to do something in most verbs, I added a search box so it can be filtered.

Some verbs which do not work well with that (such as Delete) were duplicated to have a version with search (for direct command usage through chat) and a version in the right-click interact menu.