feat: Enforce bploss on Deadzone for non-living, and allow Limit Break to be masterizable (#4)
* feat: Create search boxes for admin verbs

You were unable to search in pretty much all lists created by admin verbs. This adds a input() as Text so we can add a filter to the list first.

For some reason, this does not work in all verbs.

* feat: Make Limit Break masterizable

Hear me out. What if, instead of a random timer, we had a *weighted* random timer? One that grows?

* feat: Handle admin verbs better

Instead of scroll through a list of items to do something in most verbs, I added a search box so it can be filtered.

Some verbs which do not work well with that (such as Delete) were duplicated to have a version with search (for direct command usage through chat) and a version in the right-click interact menu.

* feat: Create entries in Server Control Panel

Instead of getting mad at our Coding overlords, you can now change Hell BP loss to whatever you want, and control Limit Break Mastery.

Also changed Limit break to be based on Sparring (leech), but this may not work properly.

* fix: Delete unrelated local files

How does this keep coming back smh I'm literally going blind and deaf of the fingers.