feat: Finally make Energies work
May Alan Turing bless our soul.

The reason why the datum/Energy/ was not accessible was because apparently using = new() on compile-time passed OK but didn't really do anything on the world.

By creating a NORMAL (not /list/!) var/energies for our mob, we can assign a /list/ to it when the player is created and from then work normally with any Energy we need.