feat: Update logging system
Now this is quite something. World.realtime is a very big number based on the BYOND epoch, and BYOND doesn't like big numbers.

This made it highly imprecise, so sometimes there would be tens of seconds of delay between each update. World.timeofday, in the other hand, uses the computer's clock and is a smaller, more precise number.

This commit covers almost all usages of it in the logging system, and also standardizes logs for admins and players.